fix(datepicker): refactor datepicker related components
