+++ packages/oruga-next/src/components/datepicker/useDatepickerMixins.ts
@@ -2,7 +2,7 @@ import { computed } from "vue";
 import { matchWithGroups } from "./utils";
 import type { DatepickerProps } from "./types";
 
-export function useDatepickerMixin(props: DatepickerProps) {
+export function useDatepickerMixins(props: DatepickerProps) {
     /**
      * Check that selected date is within earliest/latest params and
      * is within a given month
