+++ packages/oruga-next/src/components/timepicker/types.ts
@@ -1,0 +1,4 @@
+import type { ComponentProps } from "vue-component-type-helpers";
+import Timepicker from "./Timepicker.vue";
+
+export type TimepickerProps = ComponentProps<typeof Timepicker>;
